Breaking change when attr changes.
For a change to an element attribute there used to be a call for
every attribute that changed (per message).

Now, the changes are bundled together such that each message from
the Elk panel will only cause one callback to happen. The callback
instead of passing (attribute_name, value) as params now passes
a list of (attribute_name, value).

This is useful so users of the lib, such as hass, can decide how to
update their logbook, etc. It used to be that a single Elk message
would cause multiple logbook entries and the HA integration could do
nothing about that. Now there is only a single update.