Move description handler to elements.py
Bringing the code closer to where it's responsibility sits.

It was originally written as a part of util.py and then moved into
elk.py when support for multiple ElkM1 systems was added. It did fix the
multiple ElkM1 systems problem, but made the code a little less clean.

Moving it to elements.py brings the code to where the responsibilities
for descriptions should lie. This does bring a small performance hit
(really, really tiny) in that the sd_handler is called for each type
of description (panel, zones, users, counters, ...) where as in the past
the sd_handler was called exactly once across the system.

sd_handler is only called on sync, and sync in a running system rarely
happens, so the performance hit is not an impact.