Heartbeat improvements.
Moved the heartbeat code closer to where the responsibility for
connection health is -- in the proto.py module. The code is cleaner
and easier to understand. Side benefit is that the elk.py module is
starting to look cleaner and fit to its purpose of being the thing
the ties all the elements together and a high level API.