Move username checking to message.py
The username/password checking has been moved into the message module as
that is the place where as decode processing is done. There are now
"special" callbacks for login_success and login_failed. These are used
in the main Elk class.

A couple of test were fixed from a previous commit.

A couple of lint errors were cleaned up.