Clean code: moving responsibilites closer to the right spot (#31)
* Misc cleanups.

Sync handling simplied.
Memory leak in panel.py when reconnecting (adds handlers every
reconnect).
Tidyed code and comments a bit.

* Move description handler to elements.py

Bringing the code closer to where it's responsibility sits.

It was originally written as a part of util.py and then moved into
elk.py when support for multiple ElkM1 systems was added. It did fix the
multiple ElkM1 systems problem, but made the code a little less clean.

Moving it to elements.py brings the code to where the responsibilities
for descriptions should lie. This does bring a small performance hit
(really, really tiny) in that the sd_handler is called for each type
of description (panel, zones, users, counters, ...) where as in the past
the sd_handler was called exactly once across the system.

sd_handler is only called on sync, and sync in a running system rarely
happens, so the performance hit is not an impact.

* Refactor timeout names. No functional changes.

* Heartbeat improvements.

Moved the heartbeat code closer to where the responsibility for
connection health is -- in the proto.py module. The code is cleaner
and easier to understand. Side benefit is that the elk.py module is
starting to look cleaner and fit to its purpose of being the thing
the ties all the elements together and a high level API.

* Refactor rename callbk->callback.

* Refactor names plus remote disconnect_requested flag.

Disconnect requested logic simplified by using the connection_retry_time
flag. When connection_retry_time is <= 0 that implies that a reconnect
timer should not be started.

Simplified the logic when a disconnect is requested.

* Fix syntax.

* Move username checking to message.py

The username/password checking has been moved into the message module as
that is the place where as decode processing is done. There are now
"special" callbacks for login_success and login_failed. These are used
in the main Elk class.

A couple of test were fixed from a previous commit.

A couple of lint errors were cleaned up.

* Fix lint error.

Not sure that this really had to be done, but importing inside a method
is frowned upon, and, the import is linted as circular (even though
Python 3.5 handles circular imports). So, I got rid of the import!

* Refactor name. Cancel heartbeat timer just in case.