Decouple connection stuff from Elk class
The Elk class was overloaded with application layer code and connection
code. This commit splits the connection code out into a new module
called `connection.py`. The existing `proto.py` is now moved into the
new `connection.py` and make "private".

The goal in decoupling is to give a cleaner separation of concerns.

There should not be any breaking changes. API compatability was
preserved.