Refactor to decouple elements from Elk main class.
All of the elements (Areas, Zones, etc) have been refactored to not
depend on the Elk class. Part 1 of the the refactor (in the previous
commit) separated out the Connection code into its own module. This
refactor takes advantage of that for each of the elements.

For the most part this refactor is straight forward. Instead of creating
the elements to take an Elk instance they are now created taking an
Connection instance.