Separation of concerns (#53)
* Decouple connection stuff from Elk class

The Elk class was overloaded with application layer code and connection
code. This commit splits the connection code out into a new module
called `connection.py`. The existing `proto.py` is now moved into the
new `connection.py` and make "private".

The goal in decoupling is to give a cleaner separation of concerns.

There should not be any breaking changes. API compatability was
preserved.

* Refactor to decouple elements from Elk main class.

All of the elements (Areas, Zones, etc) have been refactored to not
depend on the Elk class. Part 1 of the the refactor (in the previous
commit) separated out the Connection code into its own module. This
refactor takes advantage of that for each of the elements.

For the most part this refactor is straight forward. Instead of creating
the elements to take an Elk instance they are now created taking an
Connection instance.

* Tweak CI to check push to any branch.

* Fix typing.