Message decode and events refactor (#56)
* Decouple config from connection.

Only access to url is needed in Connection, so url is now passed in
instead of entire config to Elk.

* Decouple decode from calling handlers.

Part 2 of decoupling. message.py was calling handlers meaning the
message.py was multi-purpose. If an effort to put the right logic in the
right place, message.py now just just plain ole decode. Next step will
be to pull the handler logic out.

* Remove handler code from message.py

In this part of the decouple the handler logic is separated into it's
own module. Now message.py is simply encoders and decoders. No Decode
class remains.

* Move EventHandling creation

EventHandling creation is moved out of connection and into Elk class.

* Refactor to a observer pattern.

Using Notifier class as observer of for received message and connection
status events.

* Linting.

* isort

* Fix tests from refactor.

* Add notify tests, cleanup tests.

* isort/black/lint

* isort yet again

* Fix up bin programs.

* Fix mkdoc after refactor.

* Rename to observer pattern for private vars/methods.

* Prep for release of refactor.